+++ src/components/atoms/input/RadioPriority.jsx
@@ -1,6 +1,8 @@
 import { Radio, RadioGroup, Text, Stack } from '@chakra-ui/react'
 
-const RadioPriority = () => {
+const RadioPriority = (props) => {
+  const { setPriority } = props
+
   return (
     <>
       <Text fontSize="xl" fontWeight="700">
@@ -8,9 +10,15 @@ const RadioPriority = () => {
       </Text>
       <RadioGroup fontWeight="700">
         <Stack direction="row" spacing={8}>
-          <Radio value="High">High</Radio>
-          <Radio value="Middle">Middle</Radio>
-          <Radio value="Low">Low</Radio>
+          <Radio value="High" onChange={(e) => setPriority(e.target.value)}>
+            High
+          </Radio>
+          <Radio value="Middle" onChange={(e) => setPriority(e.target.value)}>
+            Middle
+          </Radio>
+          <Radio value="Low" onChange={(e) => setPriority(e.target.value)}>
+            Low
+          </Radio>
         </Stack>
       </RadioGroup>
     </>
