+++ src/components/atoms/input/Title.jsx
@@ -1,7 +1,9 @@
 import { Input, Text } from '@chakra-ui/react'
 import React from 'react'
 
-const Title = () => {
+export default function Title(props) {
+  const { title, setTitle } = props
+
   return (
     <>
       <Text fontSize="xl" fontWeight="700">
@@ -14,9 +16,9 @@ const Title = () => {
         border="1px solid rgba(0,0,0,0.8)"
         py="32px"
         fontWeight="700"
+        value={title}
+        onChange={(e) => setTitle(e.target.value)}
       />
     </>
   )
 }
-
-export default Title
